Moves Throwable Handling to io.kotlintest package
As discussed in the pull request, there won't be any problem to keep these functions in the `io.kotlintest` package, as they're not too many. This also will avoid issues with the user, as they won't have to change anything for the bugs to be fixed and the code to still work.
