+++ kotlintest-assertions/src/main/kotlin/io/kotlintest/StrictThrowableHandling.kt
@@ -1,6 +1,4 @@
-package io.kotlintest.throwablehandling
-
-import io.kotlintest.Failures
+package io.kotlintest
 
 
 /**
